Avoid expanding arrays if not actually needed
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/lb/branches/issue-586-engine@11811 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/notifications.h
@@ -109,7 +109,7 @@ typedef struct {
   // (may be NULL)
   void **to_free;
   int to_free_length;
-  size_t to_free_size; // Allocated length
+  int to_free_size; // Allocated length
 } adlb_notif_t;
 
 /*
